Move Location type definition
